fixed Some bugs

+++ util/file.js
@@ -1,11 +1,13 @@
 const fs = require('fs');
 
 const deletFile = (filePath) => {
-    fs.unlink(filePath , (err) => {
-        if(err) {
-            throw (err);
-        }
-    });
+    if (fs.existsSync()) {
+        fs.unlink(filePath , (err) => {
+            if(err) {
+                throw (err);
+            }
+        });
+    }
 }
 
 exports.deletFile = deletFile;
